Implemented backend changes in the formula median part

+++ app/db/models.py
@@ -2,7 +2,7 @@ import uuid
 from datetime import datetime
 from typing import List, Optional
 
-from sqlalchemy import DateTime, ForeignKey, Integer, LargeBinary, String, Text, UniqueConstraint, func
+from sqlalchemy import DateTime, Float, ForeignKey, Integer, LargeBinary, String, Text, UniqueConstraint, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -64,7 +64,7 @@ class EvaluationVersion(Base):
     error: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
 
     result_json: Mapped[Optional[dict]] = mapped_column(JSONB, nullable=True)
-    final_score: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    final_score: Mapped[Optional[float]] = mapped_column(Float, nullable=True)
     grade_band: Mapped[Optional[str]] = mapped_column(String, nullable=True)
 
     created_at: Mapped[datetime] = mapped_column(
